Edit

+++ CodeWizards/src/main/java/mini/CodeWizards/repository/ChallengesRepository.java
@@ -4,5 +4,7 @@ package mini.CodeWizards.repository;
 import mini.CodeWizards.model.Challenges;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChallengesRepository extends JpaRepository<Challenges, Long> {
+import java.util.Optional;
+
+public interface ChallengesRepository extends JpaRepository<Challenges, Integer> {
 }
